backend/requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.7
Pillow>=10.0.1
django-oauth-toolkit>=1.7.1
social-auth-app-django>=5.4.0
requests>=2.31.0
reportlab>=4.0.4
celery>=5.3.4
redis>=5.0.1
openai>=1.3.0
gunicorn>=21.2.0